+++ PlantFactory.java
@@ -1,6 +1,6 @@
 public class PlantFactory {
     
-    public static Plant getPlant(String type, int age, int temperature, Coords coords){
+    public static Plant getPlant(String type, int age, double temperature, Coords coords){
         if(type.equals("Creep")){
             return(new Creep(age, temperature, coords));
         }
