+++ Water.java
@@ -2,7 +2,7 @@ import java.awt.Color;
 
 public class Water extends Liquid {
     
-    public Water(int weight, int amount, int temperature, Coords coords, Physical container){
+    public Water(double weight, int amount, double temperature, Coords coords, Physical container){
         super(temperature,weight,amount,coords,container);
     }
